backend/pyproject.toml
[project]
name = "fastApiReactTemplateBackend"
version = "0.1.0"
description = ""
requires-python = ">=3.12,<4.0"

dependencies = [
    "annotated-types==0.7.0",
    "anyio==4.10.0",
    "argon2-cffi>=25.1.0",
    "certifi==2025.8.3",
    "cffi==1.17.1",
    "colorama==0.4.6",
    "cryptography==45.0.6",
    "dnspython==2.7.0",
    "email-validator==2.2.0",
    "fastapi[standard]>=0.116.1",
    "greenlet==3.2.4",
    "idna==3.10",
    "iniconfig==2.1.0",
    "packaging==25.0",
    "passlib==1.7.4",
    "pluggy==1.6.0",
    "psycopg==3.2.6",
    "psycopg-binary==3.2.6",
    "pycparser==2.22",
    "pydantic==2.11.7",
    "pydantic-core==2.33.2",
    "pydantic-settings==2.10.1",
    "pygments==2.19.2",
    "pyjwt>=2.10.1",
    "pytest==8.4.1",
    "python-dotenv==1.1.1",
    "sentry-sdk==2.34.1",
    "sniffio==1.3.1",
    "sqlalchemy==2.0.42",
    "sqlmodel==0.0.24",
    "starlette==0.47.2",
    "tenacity==9.1.2",
    "typing-extensions==4.14.1",
    "typing-inspection==0.4.1",
    "tzdata==2025.2",
    "urllib3==2.5.0",
    "uv==0.8.8",
    "watchfiles>=1.1.0",
]

[tool.uv]
dev-dependencies = [
    "pytest==8.4.1",
    "mypy==1.17.1",
    "pre-commit==4.3.0",
    "types-passlib==1.7.7.20250408",
    "coverage==7.10.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.mypy]
strict = true
exclude = ["venv", ".venv"]

[tool.hatch.build.targets.wheel]
packages = ["fastApiReactTemplateBackend"]
